Newly arranged files
Function and required headers created to unify and making accessing socket function easier.
