+++ socket in c.layout
@@ -1,8 +1,18 @@
 [Editors]
-Order=0
-Focused=0
+Order=0,1
+Focused=1
 [Editor_0]
-CursorCol=98
-CursorRow=105
-TopLine=92
+CursorCol=16
+CursorRow=2
+TopLine=1
+LeftChar=1
+[Editor_1]
+CursorCol=30
+CursorRow=65
+TopLine=51
+LeftChar=1
+[Editor_2]
+CursorCol=44
+CursorRow=5
+TopLine=1
 LeftChar=1
